Rename Interpolator -> Interpolate
