+++ common/test/src/el/tst_Interpolate.cpp
@@ -18,7 +18,7 @@
  */
 
 #include "el/EvaluationContext.h"
-#include "el/Interpolator.h"
+#include "el/Interpolate.h"
 #include "el/Value.h"
 
 #include <string>
